make terms and conditions consistent, so remove the 2 differents component and make it into 1 terms and conditions component check if there is terms and conditions enabled, if it is then get the new terms and conditions else show the old terms and conditions
