+++ src/app/components/index.ts
@@ -1,14 +1,14 @@
-import { HmctsGlobalHeaderComponent } from './hmcts-global-header/hmcts-global-header.component';
-import { HmctsGlobalFooterComponent } from './hmcts-global-footer/hmcts-global-footer.component';
 import {
   HmctsPrimaryNavigationComponent
 } from '../../../projects/gov-ui/src/lib/components/hmcts-primary-navigation/hmcts-primary-navigation.component';
-import { CookiePolicyComponent } from './cookie-policy/cookie-policy.component';
-import { PrivacyPolicyComponent } from './privacy-policy/privacy-policy.component';
 import { AccessibilityComponent } from './accessibility/accessibility.component';
-import { ServiceDownComponent } from './service-down/service-down.component';
+import { CookiePolicyComponent } from './cookie-policy/cookie-policy.component';
 import { GetHelpComponent } from './get-help/get-help.component';
-import { LegacyTermsAndConditionsComponent } from './legacy-terms-and-conditions/legacy-terms-and-conditions.component';
+import { HmctsGlobalFooterComponent } from './hmcts-global-footer/hmcts-global-footer.component';
+import { HmctsGlobalHeaderComponent } from './hmcts-global-header/hmcts-global-header.component';
+import { PrivacyPolicyComponent } from './privacy-policy/privacy-policy.component';
+import { ServiceDownComponent } from './service-down/service-down.component';
+import { TermsAndConditionsComponent } from './terms-and-conditions/terms-and-conditions.component';
 
 export const components: any[] = [
   HmctsGlobalHeaderComponent,
@@ -19,7 +19,7 @@ export const components: any[] = [
   AccessibilityComponent,
   ServiceDownComponent,
   GetHelpComponent,
-  LegacyTermsAndConditionsComponent
+  TermsAndConditionsComponent
 ];
 
 export * from './hmcts-global-header/hmcts-global-header.component';
@@ -30,4 +30,4 @@ export * from './privacy-policy/privacy-policy.component';
 export * from './accessibility/accessibility.component';
 export * from './service-down/service-down.component';
 export * from './get-help/get-help.component';
-export * from './legacy-terms-and-conditions/legacy-terms-and-conditions.component';
+export * from './terms-and-conditions/terms-and-conditions.component';
