--- src/app/components/legacy-terms-and-conditions/legacy-terms-and-conditions.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-legacy-terms-and-conditions',
-  templateUrl: './legacy-terms-and-conditions.component.html'
-})
-export class LegacyTermsAndConditionsComponent {
-  constructor() {
-  }
-}
