+++ src/app/components/terms-and-conditions/terms-and-conditions.component.spec.ts
