+++ src/app/containers/index.ts
@@ -2,16 +2,12 @@
 import { FooterComponent } from './footer/footer.component';
 import { HeaderComponent } from './header/header.component';
 import { RedirectComponent } from './redirect/redirect.component';
-import { TermsAndConditionsComponent } from './terms-and-conditions/terms-and-conditions.component';
 
 export const containers: any[] = [
   HeaderComponent,
   FooterComponent,
-  RedirectComponent,
-  TermsAndConditionsComponent
+  RedirectComponent
 ];
 
 export * from './footer/footer.component';
 export * from './header/header.component';
-export * from './terms-and-conditions/terms-and-conditions.component';
-
