--- src/app/containers/terms-and-conditions/terms-and-conditions.component.html
@@ -1,4 +1,0 @@
-<ng-container *ngIf="document; else loading">
-  <xuilib-terms-and-conditions [document]="document"></xuilib-terms-and-conditions>
-</ng-container>
-<ng-template #loading>Loading...</ng-template>
