+++ src/register/components/check-your-answers/check-your-answers.component.html
@@ -94,7 +94,7 @@
 
   <ul class="govuk-list govuk-list--bullet">
     <li>Your personal and organisation details are correct</li>
-    <li>You have read and understood the <a routerLink="{{termsAndConditionsUrl$ | async }}" rel="noopener noreferrer" target="_blank">terms and conditions</a> and are eligible for an account to manage your organisation and cases</li>
+    <li>You have read and understood the <a routerLink="/terms-and-conditions" rel="noopener noreferrer" target="_blank">terms and conditions</a> and are eligible for an account to manage your organisation and cases</li>
     <li>You are creating an account for your organisation for the first time</li>
     <li>You will be asked to accept the terms and conditions again when logging into a MyHMCTS application for the first time</li>
   </ul>
